% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{cluster_cells}
\alias{cluster_cells}
\title{Cluster Single Cell}
\usage{
cluster_cells(
  input,
  lem,
  dims,
  method,
  embedding = NULL,
  num_clust = NULL,
  name = "Cluster",
  s = 2
)
}
\arguments{
\item{input}{the input ex_sc}

\item{lem}{the LEM within ReducedDims()}

\item{dims}{either "2d" or "Comp"}

\item{method}{can either be "spectral" or "density" which is on 2d}

\item{embedding}{if method is 2D, an embedding must be provided}

\item{num_clust}{the number of clusters. Required for spectral but optional for density.}

\item{name}{name of the colData cluster column}

\item{s}{the number of standard deviations from the curve to select cluster centers}
}
\description{
This will perform clustering on your single cell data.
}
\details{
This will perform clustering on either the high dimensional PCA / ICA components if dimension = Comp,
or the 2d tsne result if method = density. Typically spectral clustering works much better on higher dimensional data,
which density based clustering works better on 2d data.
}
\examples{
ex_sc_example <- cluster_sc(input = ex_sc_example, dimension = "Comp", method = "spectral", num_clust = 6)
}
