% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_sce.R
\name{construct_sce}
\alias{construct_sce}
\title{Construct Single Cell Experiment Object}
\usage{
construct_sce(input)
}
\arguments{
\item{input}{the input data matrix. Provide a sparse matrix}
}
\description{
This function will take an input sparse expression matrix and make a Single Cell Experiment Object
}
\details{
This will take an input matrix and create the Single Cell Experiment Object for further analysis
}
\examples{
sce <- construct_sce(input = sc_dat)
}
