% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce.R
\name{dim_reduce}
\alias{dim_reduce}
\title{Dimension Reduction}
\usage{
dim_reduce(
  input,
  genelist,
  pre_reduce,
  assay = NULL,
  nComp = 15,
  nVar = 0.85,
  log = F,
  scale = T,
  reducedDim_key = NULL,
  seed = 100
)
}
\arguments{
\item{input}{the input sce}

\item{genelist}{the subset of genes to perform dimensionality reduction on}

\item{pre_reduce}{the algorithm choice for reduction before tSNE (either "ICA", "PCA", "iPCA", or FALSE if you want to reuse).}

\item{assay}{The assay to operate on. Will default to get_def_assay(input)}

\item{nComp}{the number of components to reduce too before tSNE, 5-20 recommended.}

\item{nVar}{cutoff for percent of variance explained from PCs}

\item{log}{Whether or not to log  the input assay}

\item{scale}{Whether or not to scale the input assay}

\item{reducedDim_key}{The name of the LEM within ReducedDim slot of SCE}

\item{seed}{For tSNE, the seed. Can set to NULL if desired.}
}
\description{
This function will do dimensionality reduction.
}
\details{
If the method is ICA, independent component analysis will be performed, and then tSNE will do the final dimension reduction. If PCA is selected, PCA will be performed before on the expression matrix transpose before tSNE. This PCA will use the cells positions on the principal components. If iPCA is selected, PCA will be be performed but without transposing the data. This will create "meta cells" instead of meta genes created in the typical PCA. Then tSNE will be performed on each cells contribution (loading) to the meta cell. We find that iPCA is much more robust and leads to cleaner clusters than traditional PCA.
}
\examples{
ex_sc_example <- dim_reduce(input = ex_sc_example, genelist = gene_subset, pre_reduce = "iPCA", nComp = 15, tSNE_perp = 30, iterations = 500, print_progress=TRUE)

}
