% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{embedding}
\alias{embedding}
\title{Get embeddings}
\usage{
embedding(input, lem, embedding)
}
\arguments{
\item{input}{the input sce}

\item{the}{embedding within the LEM}
}
\description{
Returns the 2D matrix associated with a specific embedding
}
\details{

}
\examples{
sce <- get_def_assay(sce)
}
