% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embeddings}
\alias{embeddings}
\title{embeddings}
\usage{
embeddings(input)
}
\arguments{
\item{input}{the input sce}
}
\description{
A matrix of available LEM and associated embeddings
}
\details{

}
\examples{
sce <- get_def_assay(sce)
}
