% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_def_assay.R
\name{get_def_assay}
\alias{get_def_assay}
\title{Get Default Assay}
\usage{
get_def_assay(input)
}
\arguments{
\item{input}{the input sce}
}
\description{
This will get the default assay to operate on for downstream analysis.
}
\details{

}
\examples{
sce <- get_def_assay(sce)
}
