% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_genes.R
\name{get_var_genes}
\alias{get_var_genes}
\title{Get Variable Genes}
\usage{
get_var_genes(input, method, cutoff)
}
\arguments{
\item{input}{the input sce}

\item{method}{One of "CV", "Malhanobis", or "Gini}

\item{cutoff}{The threshold below which a gene will not be returned. Expressed as decimal.}
}
\description{
This will get the variable genes above a threshold, based on quantile.
}
\details{

}
\examples{
sce <- get_def_assay(sce)
}
