% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_metadata.R
\name{plot_tsne_metadata}
\alias{plot_tsne_metadata}
\title{tSNE Plot on metadata}
\usage{
plot_tsne_metadata(
  input,
  color_by,
  facet_by = NULL,
  coords = NULL,
  title = NULL,
  colors = NULL,
  size = 1.5,
  ncol = 2,
  legend_dot_size = 1.5,
  text_sizes = c(20, 10, 5, 10, 5, 5),
  shuffle = T,
  xlab = "x",
  ylab = "y"
)
}
\arguments{
\item{input}{the input ex_sc.}

\item{color_by}{What to color points by, either "UMI_sum", or pData categorial variable, ignored if gene is provided}

\item{facet_by}{What to break the plots by}

\item{coords}{A vector of character lem, x, and u. See set_xy function.
#' @param title The title}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{size}{The size of the points}

\item{ncol}{How many columns if faceting}

\item{legend_dot_size}{Size of dot in legend}

\item{text_sizes}{the text sizes on the plot}

\item{xlab}{character x label}

\item{ylab}{character y label}

\item{theme}{the ggplot theme}

\item{whether}{or not to shuffle the matrix to prevent last plotted factor blocking rest}
}
\description{
This will plot information onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(ex_sc_example, color_by = "UMI_sum", title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
