% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_def_assay.R
\name{set_def_assay}
\alias{set_def_assay}
\title{Set Default Assay}
\usage{
set_def_assay(input, def_assay)
}
\arguments{
\item{input}{the input sce}

\item{assay}{the character name of the default assay}
}
\description{
This will set and get the default assay to operate on for downstream analysis.
}
\details{

}
\examples{
sce <- set_def_assay(sce, "counts")
}
