% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_xy.R
\name{set_xy}
\alias{set_xy}
\title{Set xy}
\usage{
set_xy(input, lem, x = "x", y = "y")
}
\arguments{
\item{input}{the input sce}

\item{lem}{the LEM within reducedDims}

\item{x}{The x values for 2D representation. Can be character x or a numeric column of lem@sampleFactors}

\item{y}{The y values for 2D representation. Can be character y or a numeric column of lem@sampleFactors}
}
\description{
This will set default X and Y coordinates based on the LEM results within reducedDims(sce)
}
\details{

}
\examples{
sce <- set_xy(input = sce, lem = "iPCA", x = "x", y = "y")
}
