% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lib_size.R
\name{calc_libsize}
\alias{calc_libsize}
\title{Calculate Library Size}
\usage{
calc_libsize(input, assay = "counts", label = NULL)
}
\arguments{
\item{input}{the input sce object}

\item{assay}{which assay to calculate library size on}

\item{suffix}{the}
}
\description{
This function will count the number of reads / UMIs per cell
}
\details{
This will calculate the total number of UMIs on a per cell basis.
}
\examples{
sce <- calc_libsize(input = sce)
}
